ai-service/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
torch==2.5.1
torchvision==0.20.1
Pillow==11.2.1
python-multipart==0.0.20
python-dotenv==1.1.0
numpy==2.2.5
grad-cam==1.5.5
opencv-python-headless==4.13.0.92
transformers==5.5.4
